Convert SwFrameFormats to boost::multi_index
This is almost the same situation as SwPageDescs. What makes this
more complicated is the fact, that actually duplicated draw objects
are allowed, in regard to the key values "type" and "name".

And actually for some types, duplicate names are not allowed, e.g.
SwDoc::FindFlyByName( const OUString& rName, sal_Int8 nNdTyp )
expects a single result!

Change-Id: I6e0ea1099c1c1e6cfe90926170e27179722e88b8