Don't reschedule SFX bindings on timeout change
Drops stopping the tasks before changing the timeout, otherwise the
task is recreated and put to the front of the scheduling queue as a
new task.

This also allows us to drop the special SFX append-to-task-list
handling.

For readability the patch also renames the job timer, and converts
the dispatch timer to an idle.

Change-Id: Ida0c412e71babb431bdfb360abfa9ad1109d4fd7