Use InsertPage in AssertFlyPages
AssertFlyPages used it's own code to insert pages. It had two
flaws:

  1. It used the SwPageDesc of the previous page, unaware of
     changed page description of a new page, which would later
     be fixed by CheckPageDescs, eventually requiring re-layout.

  2. The Code has an of-by-one error when setting the initial
     bOdd, which would also insert wrong / inverse left and
     right pages.

So this changes AssertFlyPages to use InsertPage, which has a
little more overhead for inserts, but prevents later expensive
corrections, especially when CheckPageDescs() inserts or removes
"early" pages, which need to evaluate and move all page bound
flys.

Change-Id: I8806cebb26db60602d8438f117e0416e07228b3b