Reorganize Scheduler priority classes
This is based on glibs classification of tasks, but while glib uses
an int for more fine grained priority, we stay with our enum.

1. Timers start with DEFAULT priority, which directly corresponds
   with the previous HIGH priority
2. Idles start with DEFAULT_IDLE priority instead of the previous
   HIGH priority, so idle default becomes "really run when idle".

As RESIZE and REPAINT are special, and the DEFAULTS are set, there
is just one primary decision for the programmer: should my idle
run before paint (AKA HIGH_IDLE)?

And this includes the defintion to skip idles with
priority >= DEFAULT_IDLE in busy times. This should proably be
expanded to also include timers, as priority is already handled
independent from the task type.

If we really need a more fine-grained classification, we can add it
later, or also switch to a real int. As a result, this drops many
classifications from the code and drastically changes behaviour,
AKA a mail merge from KDE is now as fast as Gtk+ again.

Change-Id: I498a73fd02d5fb6f5d7e9f742f3bce972de9b1f9